base64 and crc cleanups

+++ include/ptpgp/crc24.h
@@ -1,7 +1,7 @@
 typedef struct {
   ptpgp_err_t last_err;
   bool done;
-  long crc;
+  uint32_t crc;
 } ptpgp_crc24_t;
 
 ptpgp_err_t ptpgp_crc24_init(ptpgp_crc24_t *);
